Move `getUsernamesWithRoles` to utils
