+++ src/utils/tableRequestStatusIndex.ts
@@ -14,7 +14,7 @@ import {
 // Unfortunately the order of original enums from `schema.graphql` are lost
 // when types are generated for the frontend in `global-types`.
 // Generated enum strings are maintained in such a way here for react table column sorting.
-function parseSortIndex(arr: string[]): () => { [v: string]: number } {
+export function parseSortIndex(arr: string[]): () => { [v: string]: number } {
   return once(() => Object.fromEntries(arr.map((v, i) => [v, i])));
 }
